feat: embed PHP apps into the FrankenPHP binary

+++ build-static.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
 
 set -o errexit
-set -o xtrace
 
 if ! type "git" > /dev/null; then
     echo "The \"git\" command must be installed."
@@ -48,37 +47,47 @@ fi
 
 bin="frankenphp-$os-$arch"
 
-mkdir -p dist/
-cd dist/
-
-if [ -d "static-php-cli/" ]; then
-    cd static-php-cli/
-    git pull
-else
-    git clone --depth 1 https://github.com/crazywhalecc/static-php-cli
-    cd static-php-cli/
+if [ "$CLEAN" ]; then
+    rm -Rf dist/
+    go clean -cache
 fi
 
-composer install --no-dev -a
+# Build libphp if ncessary
+if [ -f "dist/static-php-cli/buildroot/lib/libphp.a" ]; then
+    cd dist/static-php-cli    
+else
+    mkdir -p dist/
+    cd dist/
 
-
-if type "brew" > /dev/null; then
-    packages="composer"
-    if [ "$RELEASE" ]; then
-        packages="$packages gh"
+    if [ -d "static-php-cli/" ]; then
+        cd static-php-cli/
+        git pull
+    else
+        git clone --depth 1 https://github.com/crazywhalecc/static-php-cli
+        cd static-php-cli/
     fi
 
-    brew install --formula --quiet "$packages"
+    if type "brew" > /dev/null; then
+        packages="composer"
+        if [ "$RELEASE" ]; then
+            packages="$packages gh"
+        fi
+
+        brew install --formula --quiet "$packages"
+    fi
+
+    composer install --no-dev -a
+
+    if [ "$os" = "linux" ]; then
+        extraOpts="--disable-opcache-jit"
+    fi
+
+    ./bin/spc doctor
+    ./bin/spc fetch --with-php="$PHP_VERSION" --for-extensions="$PHP_EXTENSIONS"
+    # shellcheck disable=SC2086
+    ./bin/spc build --enable-zts --build-embed $extraOpts "$PHP_EXTENSIONS" --with-libs="$PHP_EXTENSION_LIBS"
 fi
 
-if [ "$os" = "linux" ]; then
-    extraOpts="--disable-opcache-jit"
-fi
-
-./bin/spc doctor
-./bin/spc fetch --with-php="$PHP_VERSION" --for-extensions="$PHP_EXTENSIONS"
-# shellcheck disable=SC2086
-./bin/spc build --enable-zts --build-embed $extraOpts "$PHP_EXTENSIONS" --with-libs="$PHP_EXTENSION_LIBS"
 CGO_CFLAGS="-DFRANKENPHP_VERSION=$FRANKENPHP_VERSION $(./buildroot/bin/php-config --includes | sed s#-I/#-I"$PWD"/buildroot/#g)"
 export CGO_CFLAGS
 
@@ -92,15 +101,28 @@ export CGO_LDFLAGS
 LIBPHP_VERSION="$(./buildroot/bin/php-config --version)"
 export LIBPHP_VERSION
 
-cd ../../caddy/frankenphp/
+cd ../..
+
+# Embed PHP app, if any
+if [ -d "$EMBED" ]; then
+    mv embed embed.bak
+    cp -R "$EMBED" embed
+fi
+
+cd caddy/frankenphp/
 go env
 go build -buildmode=pie -tags "cgo netgo osusergo static_build" -ldflags "-linkmode=external -extldflags -static-pie -w -s -X 'github.com/caddyserver/caddy/v2.CustomVersion=FrankenPHP $FRANKENPHP_VERSION PHP $LIBPHP_VERSION Caddy'" -o "../../dist/$bin"
+cd ../..
 
-cd ../../dist/
-"./$bin" version
+if [ -d "$EMBED" ]; then
+    rm -Rf embed
+    mv embed.bak embed
+fi
+
+"dist/$bin" version
 
 if [ "$RELEASE" ]; then
-    gh release upload "$FRANKENPHP_VERSION" "$bin" --repo dunglas/frankenphp --clobber
+    gh release upload "$FRANKENPHP_VERSION" "dist/$bin" --repo dunglas/frankenphp --clobber
 fi
 
 if [ "$CURRENT_REF" ]; then
